wip: refactor folder structure, add shared packages and helpers.
