+++ pkg/data/.gothicCli/templates/Dockerfile-template
