+++ pkg/data/.gothicCli/templates/samconfig-template.toml
