+++ pkg/data/.gothicCli/templates/template-custom-domain-with-arn.yaml
