+++ pkg/data/.gothicCli/templates/template-custom-domain.yaml
