+++ pkg/data/.gothicCli/templates/template-default.yaml
