+++ pkg/data/gothic-config-full-sample.json
