+++ pkg/data/gothic-config.json
