+++ pkg/data/public/favicon.ico
