+++ pkg/data/src/components/optimizedImage.templ
