+++ pkg/data/tailwind.config.js
